% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContriBOT-package.R
\docType{package}
\name{ContriBOT-package}
\alias{ContriBOT}
\alias{ContriBOT-package}
\title{ContriBOT: Detection and extraction of Author Contribution statements, Acknowledgements, and ORCIDs from scientific papers}
\description{
Detection and extraction of Author Contribution statements, Acknowledgements, and ORCIDs from scientific papers.
}
\author{
\strong{Maintainer}: Vladislav Nachev \email{vladislav.nachev@charite.de} (\href{https://orcid.org/0000-0003-0521-6153}{ORCID})

Authors:
\itemize{
  \item Fyeqa Akram
}

}
\keyword{internal}
