% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContriBOT.R
\name{extract_contributions}
\alias{extract_contributions}
\title{Seach for CRediT keywords and extract section.}
\usage{
extract_contributions(pdf_text_sentences)
}
\arguments{
\item{pdf_text_sentences}{Document corpus loaded with the pdf_load function.}
}
\value{
Tibble with one row per screened document and the file name and logical values for authorship,
acknowledgement, and orcid statements detected as columns,
plus additional columns that contain the statements that were extracted.
}
\description{
The algorithm searchers for a  Authorship, Acknowledgement, and ORCID statements
and extracts them. If no known section title was found, the algorithm searches
for several categories of similar keywords in each sentence of the whole manuscript.
}
\examples{
\dontrun{
extract_contributions(pdf_load("examples/"))
}

}
