% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_extraction.R
\name{extract_orcid_hyperlink}
\alias{extract_orcid_hyperlink}
\title{Seach for ORCID hyperlinks and extract ORCIDs.}
\usage{
extract_orcid_hyperlink(PDF_file)
}
\arguments{
\item{PDF_file}{String with the path to the PDF to be screened.}
}
\value{
String with the ORCIDs
}
\description{
The algorithm searchers for an ORCID and extracts what it finds.
}
\examples{
\dontrun{
extract_orcid_hyperlink(some_pdf)
}

}
