% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_credit_roles}
\alias{parse_credit_roles}
\title{Parse an authorship statement that is list-formatted, especially as CRediT, and extract
the authors and roles into a tidy tibble, with a single author and role per row.}
\usage{
parse_credit_roles(credit_str)
}
\arguments{
\item{credit_str}{String with extracted contribution statement with the
\code{extract_contributions} function.}
}
\value{
Tibble with one row per author and role, or NA if the statement could
not be parsed, for example if narrative or formatted in a non-standardized way.
}
\description{
Parse an authorship statement that is list-formatted, especially as CRediT, and extract
the authors and roles into a tidy tibble, with a single author and role per row.
}
\examples{
\dontrun{
parse_credit_roles("<section> Author Contributions <section>:
 study concepts: A.B. and C.D; study design: A.B. and E.F;
  data acquisition: G.H, A.B., and C.D.;
   quality control of data and algorithms: A.B., G.H., I.J., and C.D.;
    data analysis and interpretation: K.L., I.J., G.H., and C.D.;
     statistical analysis: G.H., I.J., and K.L;
      manuscript preparation: A.B. and E.F.; manuscript editing: all;
       manuscript review: all.")
# or if we have a tibble of extracted and classified authorship statements:
library(dplyr)
library(purrr)
parsed_results <- credit_results |>
 pull(contrib_statement) |>
 map(parse_credit_roles)
}

}
