% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContriBOT.R
\name{extract_contributions}
\alias{extract_contributions}
\title{Search for Authorship, Acknowledgement, and ORCID keywords and extract corresponding sections.}
\usage{
extract_contributions(pdf_text_sentences)
}
\arguments{
\item{pdf_text_sentences}{Document corpus loaded with the \code{oddpub::pdf_load} function.}
}
\value{
Tibble with one row per screened document and the file name and logical values for authorship,
acknowledgement, and orcid statements detected as columns,
plus additional columns that contain the statements that were extracted.
}
\description{
The algorithm searchers for a  Authorship, Acknowledgement, and ORCID statements
and extracts them.
}
\examples{
\dontrun{
extract_contributions(pdf_load("examples/"))
}

}
